Move funções relacionadas aos territórios para um único arquivo
