+++ tasks/get_territorie_info.py
@@ -1,5 +1,34 @@
 
 import unicodedata
+from typing import Dict, Iterable
+
+from .interfaces import DatabaseInterface
+
+
+def get_territories_gazettes(
+    database: DatabaseInterface,
+) -> Iterable[Dict]:
+
+    command = """
+    SELECT 
+        *
+    FROM
+        territories
+    ;
+    """
+
+    territories = [format_territories_data(territory) for territory in database.select(command)]
+
+    return territories
+
+
+def format_territories_data(data):
+    return {
+        "id": data[0],
+        "territory_name": data[1],
+        "state_code": data[2],
+        "state": data[3],
+    }
 
 
 def get_territorie_info(state: str, name: str, territories: list):
@@ -12,7 +41,7 @@ def get_territorie_info(state: str, name: str, territories: list):
         if territorie["state"].lower() == state.lower() and territorie_name == name:
 
             return territorie["id"], territorie["territory_name"], territorie["state_code"]
-    
+
 
 def limpar_name(name: str):
